requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.10
chromadb>=0.4.22
pydantic>=2.5.0
pypdf>=3.17.0
pandas>=2.1.0
openpyxl>=3.1.2
unstructured[docx]>=0.12.0
python-multipart>=0.0.6
tiktoken>=0.5.2
aiofiles>=23.2.0
python-docx>=1.1.0
